requirements.txt
torch>=1.9.0
numpy>=1.21.0
python-chess>=0.31.4
